requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
pymongo==4.3.3
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.3
